// examin_rest_api_plus_frontend/examin-web-front-master/src/app/models/bank.question.model.ts
export class BankQuestion {
    categoryId: number;
    name: string;
    questionId: number;
    questionL1: string;
    questionL2: string;
    AL1: string;
    AL2: string;
    BL1: string;
    BL2: string;
    CL1: string;
    CL2: string;
    DL1: string;
    DL2: string;
    correctAnswer?: string;
}